requirements.txt
# Core numerical and data handling libraries
numpy
scipy
pandas
PyYAML 

# Visualization and UI
pygame
matplotlib
seaborn

# Testing
pytest